+++ changelog.md
@@ -3,7 +3,9 @@
 
 - Add an interactive/scriptable mode where messages are read from stdin
 - Re-structure exports to improve readibility of the documentation
-- Close connections in addition to sending gtfo when idle time exceeded (needs http2-client-0.3.0.2)
+- Close connections in addition to sending http2 gtfo when idle time exceeded (needs http2-client-0.3.0.2)
+- Add a closeSession method
+- Close sessions when they are garbage collected
 
 0.1.0.3
 =======